If localization not exist default, default attribute become required. Also If localization is default, it should not delete.

+++ app/Services/LocalizationService.php
@@ -110,6 +110,9 @@ class LocalizationService
     public function delete($id)
     {
         $data = $this->find($id);
+        if ($data->default) {
+            return false;
+        }
         return $data->delete();
     }
